refactor: entities index.ts 생성 및 연관되는 경로들 수정
