+++ src/entities/feed/index.ts
@@ -1,0 +1,3 @@
+// NOTE: user model
+export { FeedDetailItem } from './model/feedDetailTypes';
+export { FeedItem } from './model/feedTypes';
