+++ src/entities/recipe/index.ts
@@ -1,0 +1,10 @@
+// NOTE: recipe model
+export { Recipe } from './model';
+export { mockRecipes } from './model/mockRecipe';
+export { RecipeDetailItem } from './model/RecipeDetailTypes';
+
+// NOTE: recipe UI
+export { default as ArticleView } from './ui/ArticleView';
+export { default as DetailDeleteComponent } from './ui/DetailDeleteComponent';
+export { default as FeedView } from './ui/FeedView';
+export { default as ImageView } from './ui/ImageView';
