+++ src/entities/user/index.ts
@@ -1,0 +1,9 @@
+// NOTE: user api
+export { mockUser } from './api/mockUser';
+
+// NOTE: user model
+export { User } from './model/userType';
+
+// NOTE: user UI
+export { default as Button } from './ui/Button';
+export { default as UserHeader } from './ui/UserHeader';
