+++ src/features/auth/ui/KakaoLoginButton.tsx
@@ -3,7 +3,7 @@ import { Text } from 'react-native';
 
 import KakaoSvg from '@/assets/img/auth/kakao.svg';
 import { kakaoLogin } from '@features/auth/api/login';
-import Button from '@entities/user/ui/Button';
+import { Button } from '@entities/user';
 import { RootNavigationProp } from '@shared/types/navigation';
 
 interface Props {
