+++ src/features/feed/api/dummyFeedDetail.ts
@@ -1,4 +1,4 @@
-import { FeedDetailItem } from '@entities/feed/model/feedDetailTypes';
+import { FeedDetailItem } from '@entities/feed';
 
 export const dummyFeedDetail: FeedDetailItem = {
   id: '2',
