+++ src/features/feed/api/getDetailFeedData.ts
@@ -1,4 +1,4 @@
-import { FeedDetailItem } from '@entities/feed/model/feedDetailTypes';
+import { FeedDetailItem } from '@entities/feed';
 import { dummyFeedDetail } from './dummyFeedDetail';
 
 export const apiGetDetailData = async (id: string): Promise<FeedDetailItem> => {
