+++ src/features/feed/api/getFeedData.ts
@@ -1,4 +1,4 @@
-import { FeedItem } from '@entities/feed/model/feedTypes';
+import { FeedItem } from '@entities/feed';
 import { dummyFeeds } from './dummyFeeds';
 
 export const apiGetData = async (page: number, limit: number): Promise<FeedItem[]> => {
