+++ src/features/feed/model/useFeedData.ts
@@ -1,6 +1,6 @@
 import { useRef, useState, useEffect, useCallback } from 'react';
 
-import { FeedItem } from '@entities/feed/model/feedTypes';
+import { FeedItem } from '@entities/feed';
 import { apiGetData } from '../api/getFeedData';
 
 export const useFeedData = () => {
