+++ src/features/user/ui/header/UserHeaderSection.tsx
@@ -3,7 +3,7 @@ import { View, Text, Image } from 'react-native';
 
 import EditProfileButton from '@features/user/ui/EditProfileButton';
 import UserTabs from '@features/user/ui/UserTabs';
-import { User } from '@entities/user/model/userType';
+import { User } from '@entities/user';
 import { defaultShadow } from '@shared/ui/defaultShadow';
 
 import ProfileEditModal from '../ProfileEditModal';
