+++ src/pages/feed/ui/Feed.tsx
@@ -5,7 +5,7 @@ import { SafeAreaView } from 'react-native-safe-area-context';
 import { useFeedData } from '@features/feed/model/useFeedData';
 import { FeedCard } from '@features/feed/ui/FeedCard';
 import FeedDetailModal from '@features/feed/ui/FeedDetailModal';
-import { FeedItem } from '@entities/feed/model/feedTypes';
+import { FeedItem } from '@entities/feed';
 
 const Feed = () => {
   const { dataList, onEndReached, onRefresh, isRefreshing } = useFeedData();
